# openapi3.yaml
openapi: 3.0.1
info:
  title: ingestion-trigger
  description: >-
    Service that designed to validate and trigger ingestion of new layers from
    raw data and update existing layers
  version: 3.2.5
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
paths:
  /ingestion:
    post:
      operationId: ingestNewLayer
      tags:
        - ingestion
      summary: start a process of creating new layer from raw data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestionNewLayerRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessfulIngestionResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: >-
                  ./Schema/ingestionTrigger/responses/ingestionTriggerResponses.yaml#/components/schemas/errorMessage
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: >-
                  ./Schema/ingestionTrigger/responses/ingestionTriggerResponses.yaml#/components/schemas/errorMessage
        '422':
          description: Unprocessable Content
          content:
            application/json:
              schema:
                $ref: >-
                  ./Schema/ingestionTrigger/responses/ingestionTriggerResponses.yaml#/components/schemas/errorMessage
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: >-
                  ./Schema/ingestionTrigger/responses/ingestionTriggerResponses.yaml#/components/schemas/errorMessage
  /ingestion/{id}:
    put:
      operationId: updateLayer
      tags:
        - ingestion
      summary: start a process of updating an existing layer from raw data
      parameters:
        - name: id
          in: path
          description: The id of the layer to update
          required: true
          schema:
            $ref: ./Schema/core/layerMetadata.yaml#/components/schemas/LayerId
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestionUpdateLayerRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessfulIngestionResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: >-
                  ./Schema/ingestionTrigger/responses/ingestionTriggerResponses.yaml#/components/schemas/errorMessage
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: >-
                  ./Schema/ingestionTrigger/responses/ingestionTriggerResponses.yaml#/components/schemas/errorMessage
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: >-
                  ./Schema/ingestionTrigger/responses/ingestionTriggerResponses.yaml#/components/schemas/errorMessage
        '422':
          description: Unprocessable Content
          content:
            application/json:
              schema:
                $ref: >-
                  ./Schema/ingestionTrigger/responses/ingestionTriggerResponses.yaml#/components/schemas/errorMessage
        '500':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: >-
                  ./Schema/ingestionTrigger/responses/ingestionTriggerResponses.yaml#/components/schemas/errorMessage
  /ingestion/validate/gpkgs:
    post:
      operationId: validateGpkgs
      tags:
        - ingestion
      summary: checks that all provided gpkgs files have valid raw data
      requestBody:
        description: A json object containing array of gpkg files paths
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GpkgInputFiles'
            example:
              gpkgFilesPath: ['path/to/blue_marble.gpkg']
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: >-
                  ./Schema/ingestionTrigger/responses/ingestionTriggerResponses.yaml#/components/schemas/validateSourcesResponse
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: >-
                  ./Schema/ingestionTrigger/responses/ingestionTriggerResponses.yaml#/components/schemas/errorMessage
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: >-
                  ./Schema/ingestionTrigger/responses/ingestionTriggerResponses.yaml#/components/schemas/errorMessage
  /info/gpkgs:
    post:
      operationId: getGpkgsInfo
      tags:
        - info
      summary: returns the GDAL info for the provided gpkgs files
      requestBody:
        description: A json object containing array of gpkg files path
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GpkgInputFiles'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: >-
                  ./Schema/ingestionTrigger/responses/ingestionTriggerResponses.yaml#/components/schemas/sourcesInfoResponse
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: >-
                  ./Schema/ingestionTrigger/responses/ingestionTriggerResponses.yaml#/components/schemas/errorMessage
        '404':
          description: Files not found
          content:
            application/json:
              schema:
                $ref: >-
                  ./Schema/ingestionTrigger/responses/ingestionTriggerResponses.yaml#/components/schemas/errorMessage
        '422':
          description: Unprocessable Content
          content:
            application/json:
              schema:
                $ref: >-
                  ./Schema/ingestionTrigger/responses/ingestionTriggerResponses.yaml#/components/schemas/errorMessage
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: >-
                  ./Schema/ingestionTrigger/responses/ingestionTriggerResponses.yaml#/components/schemas/errorMessage

components:
  schemas:
    CallbackUrls:
      type: array
      minItems: 1
      items:
        type: string
        pattern: ^https?:\/\/[^\s/$.?#].[^\s]*$
        description: callback url for ingestion status notifications
      example: ['https://my-dns-for-callback']
    GpkgFilesPath:
      type: array
      minItems: 1
      maxItems: 1
      items:
        type: string
        pattern: ^(\/?[\w-]+)(\/[\w-]+)*\/[\wא-ת\.-]+\.gpkg$
        description: gpkg source file paths
      example: ['/path/to/example.gpkg']
    IngestionNewLayerMetadata:
      type: object
      required:
        - classification
        - productId
        - productName
        - productType
        - srs
        - srsName
        - transparency
        - region
      properties:
        productId:
          $ref: ./Schema/core/layerMetadata.yaml#/components/schemas/ProductId
        productName:
          $ref: ./Schema/core/layerMetadata.yaml#/components/schemas/ProductName
        productType:
          $ref: ./Schema/core/layerMetadata.yaml#/components/schemas/ProductType
        productSubType:
          $ref: ./Schema/core/layerMetadata.yaml#/components/schemas/ProductSubType
        region:
          $ref: ./Schema/core/layerMetadata.yaml#/components/schemas/Region
        srs:
          $ref: ./Schema/core/layerMetadata.yaml#/components/schemas/Srs
        srsName:
          $ref: ./Schema/core/layerMetadata.yaml#/components/schemas/SrsName
        transparency:
          $ref: ./Schema/core/layerMetadata.yaml#/components/schemas/Transparency
        producerName:
          $ref: ./Schema/core/layerMetadata.yaml#/components/schemas/ProducerName
        scale:
          $ref: ./Schema/core/layerMetadata.yaml#/components/schemas/Scale
        description:
          $ref: ./Schema/core/layerMetadata.yaml#/components/schemas/Description
        classification:
          $ref: ./Schema/core/layerMetadata.yaml#/components/schemas/Classification
    IngestionNewLayerRequest:
      type: object
      required:
        - metadata
        - inputFiles
        - ingestionResolution
      properties:
        metadata:
          $ref: '#/components/schemas/IngestionNewLayerMetadata'
        inputFiles:
          $ref: '#/components/schemas/InputFiles'
        ingestionResolution:
          $ref: '#/components/schemas/IngestionResolution'
        callbackUrls:
          $ref: '#/components/schemas/CallbackUrls'
    IngestionResolution:
      type: number
      format: double
      minimum: 0.000000167638063430786
      maximum: 0.703125
      description: requested ingestion resolution
    IngestionUpdateLayerMetadata:
      type: object
      required:
        - classification
      properties:
        classification:
          $ref: ./Schema/core/layerMetadata.yaml#/components/schemas/Classification
    IngestionUpdateLayerRequest:
      type: object
      required:
        - metadata
        - inputFiles
        - ingestionResolution
      properties:
        metadata:
          $ref: '#/components/schemas/IngestionUpdateLayerMetadata'
        inputFiles:
          $ref: '#/components/schemas/InputFiles'
        ingestionResolution:
          $ref: '#/components/schemas/IngestionResolution'
        callbackUrls:
          $ref: '#/components/schemas/CallbackUrls'
    InputFiles:
      type: object
      required:
        - gpkgFilesPath
        - metadataShapefilePath
        - productShapefilePath
      properties:
        gpkgFilesPath:
          $ref: '#/components/schemas/GpkgFilesPath'
        metadataShapefilePath:
          $ref: '#/components/schemas/MetadataShapefilePath'
        productShapefilePath:
          $ref: '#/components/schemas/ProductShapefilePath'
    MetadataShapefilePath:
      type: string
      pattern: ^(\/?[\w-]+)(\/[\w-]+)*\/ShapeMetadata\.shp$
      description: metadata shape file path
      example: 'path/to/ShapeMetadata.shp'
    ProductShapefilePath:
      type: string
      pattern: ^(\/?[\w-]+)(\/[\w-]+)*\/Product\.shp$
      description: product shape file path
      example: 'path/to/Product.shp'
    GpkgInputFiles:
      type: object
      required:
        - gpkgFilesPath
      properties:
        gpkgFilesPath:
          $ref: '#/components/schemas/GpkgFilesPath'
    SuccessfulIngestionResponse:
      type: object
      required:
        - jobId
        - taskId
      properties:
        jobId:
          type: string
          format: uuid
        taskId:
          type: string
          format: uuid